+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat(slices-intersection_by): intersection_by function added
 * ✅ test(slices-intersection): intersection test cases and benchmarks added
 * 🎉 feat(slices-intersection): intersection function added
 * ✅ test(slices-from_pairs): from_pairs test cases and benchmarks added